normalize evaluator ids and consolidate semantic evaluator
